+++ testapp/src/test/java/testapp/endpoint/GHIssue504.java
@@ -9,4 +9,10 @@ public class GHIssue504 extends EndpointTester {
         url("/gh/504").get();
         bodyEq("10");
     }
+
+    @Test
+    public void testNoDef() throws Exception {
+        url("/gh/504/no_def").get();
+        bodyEq("0");
+    }
 }