+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 **1.7.2** 
+* Running `Config` sample project in prod mode cannot find template #527
 * Large content sent to response get truncated #531
 * Extends error response to allow user defined code #530
 * Add `Controller.Util.download(URL)` helper method #529